generate data for all classes, service for all classes, can generate a car now
